Move the semantics (raw text interpretation) from Parser and Model to Handler
Motivation: the parser should parse and not interpret. User should have the freedom to interpret in his custom Handler. The same hold for the Model: the Model should be as dumb as possible holding just the data. Any interpretation should happen outside of the model - i.e. in the Handler that creates the model.