Move error handling from EditorConfigHandler to a new interface ErrorHandler
Motivation: before this change the strictness of the parser and error
handling are scattered over the Parser and Handler. First, the error
handling and strictness should be treated as the same thing in the ec4j
API. Strictness should be implemented by the behavior of the given
ErrorHandler.

Second, we believe, that the users of ec4j will mostly use the provided
Handlers (esp. the EditorConfigModelHandler) unchanged, while they will
probably have very distinct needs as for strictness and error handling.
Therefore the ErrorHandler interface should be independent of the
general content handling EditorConfigHandler.
should be able to configure the strictness of
the parser independently of the Handler implementation.